requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
idna==3.6
pipreqs==0.4.13
pyaes==1.6.1
pyasn1==0.5.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
Telethon==1.29.3
urllib3==2.1.0
yarg==0.1.9
